requirements.txt
pyodbc
pandas
scikit-learn
SQLAlchemy
python-decouple
matplotlib
scipy
seaborn